requirements.txt
# 图像处理
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0

# 文档处理  
python-docx>=0.8.11

# AI服务集成
openai>=1.0.0
requests>=2.25.0

# 环境配置
python-dotenv>=0.19.0

# 类型注解支持
typing-extensions>=4.0.0

# 多模型AI服务支持
# - OpenRouter API (主要服务)
# - SiliconFlow API (DeepSeek备用)
# - Moonshot API (Kimi标题生成)
